chore: remove dead code relating to single hop trade ui (#6117)
* fix: dont nuke trade input state when backing out of a trade preview

* fix: persist user decision to enter fiat amounts when trading

* fix: clear input amount when switching assets

* chore: remove dead code relating to single hop trade ui

* chore: move isTradingActive helper to swapper api helpers

* chore: move useApprovalTx hook into multi hop trade confirm component hooks

* chore: move approval hook helpers into multi op trade confirm hooks directory

* chore: remove unused multi hop feature flag

---------
