+++ .env.base
@@ -29,7 +29,6 @@ REACT_APP_FEATURE_WALLET_CONNECT_V2=true
 REACT_APP_FEATURE_FOX_DISCOUNTS=true
 REACT_APP_FEATURE_THORCHAIN_LENDING=true
 REACT_APP_FEATURE_THORCHAIN_LP=false
-REACT_APP_FEATURE_MULTI_HOP_TRADES=true
 REACT_APP_FEATURE_THORCHAINSWAP_LONGTAIL=true
 
 # absolute URL prefix
