+++ src/components/MultiHopTrade/MultiHopTrade.tsx
@@ -5,7 +5,6 @@ import { AnimatePresence } from 'framer-motion'
 import { lazy, memo, Suspense, useEffect } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import { MemoryRouter, Route, Switch, useLocation, useParams } from 'react-router-dom'
-import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 import { selectAssetById } from 'state/slices/assetsSlice/selectors'
 import { tradeInput } from 'state/slices/tradeInputSlice/tradeInputSlice'
 import { tradeQuoteSlice } from 'state/slices/tradeQuoteSlice/tradeQuoteSlice'
@@ -13,21 +12,12 @@ import { useAppDispatch, useAppSelector } from 'state/store'
 
 import { TradeRoutePaths } from './types'
 
-const Approval = lazy(() =>
-  import('./components/Approval/Approval').then(({ Approval }) => ({ default: Approval })),
-)
 const MultiHopTradeConfirm = lazy(() =>
   import('./components/MultiHopTradeConfirm/MultiHopTradeConfirm').then(
     ({ MultiHopTradeConfirm }) => ({
       default: MultiHopTradeConfirm,
     }),
   ),
-)
-
-const TradeConfirm = lazy(() =>
-  import('./components/TradeConfirm/TradeConfirm').then(({ TradeConfirm }) => ({
-    default: TradeConfirm,
-  })),
 )
 
 const TradeInput = lazy(() =>
@@ -41,7 +31,6 @@ const VerifyAddresses = lazy(() =>
 
 const MultiHopEntries = [
   TradeRoutePaths.Input,
-  TradeRoutePaths.Approval,
   TradeRoutePaths.Confirm,
   TradeRoutePaths.VerifyAddresses,
 ]
@@ -85,7 +74,6 @@ export const MultiHopTrade = memo(({ defaultBuyAssetId, ...cardProps }: TradeCar
 const MultiHopRoutes = memo(() => {
   const location = useLocation()
   const dispatch = useAppDispatch()
-  const enableMultiHopTrades = useFeatureFlag('MultiHopTrades')
 
   useEffect(() => {
     return () => {
@@ -105,10 +93,7 @@ const MultiHopRoutes = memo(() => {
             <TradeInput />
           </Route>
           <Route key={TradeRoutePaths.Confirm} path={TradeRoutePaths.Confirm}>
-            {enableMultiHopTrades ? <MultiHopTradeConfirm /> : <TradeConfirm />}
-          </Route>
-          <Route key={TradeRoutePaths.Approval} path={TradeRoutePaths.Approval}>
-            <Approval />
+            <MultiHopTradeConfirm />
           </Route>
           <Route key={TradeRoutePaths.VerifyAddresses} path={TradeRoutePaths.VerifyAddresses}>
             <VerifyAddresses />
